requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
dj-database-url
requests
